moved util.js to root of source directory
